requirements.txt
# Core
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
rasterio>=1.3.0
fiona>=1.9.0
pyproj>=3.6.0
contextily>=1.3.0

# Satellite Data
sentinelhub>=3.9.0
eoreader>=0.28.0
sentinelsat>=1.1.0
landsatxplore>=0.10.0

# Machine Learning
scikit-learn>=1.3.0
scikit-image>=0.22.0
opencv-python>=4.8.0
tensorflow>=2.13.0
torch>=2.0.0

# Visualization
plotly>=5.17.0
folium>=0.14.0
keplergl>=0.3.0
pydeck>=0.8.0
streamlit-folium>=0.15.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
